requirements.txt
# Core for Google AI Studio (Gemini API)
google-generativeai>=0.7.0,<1.0
python-dotenv>=1.0.1,<2.0
